+++ Application/EnemySpawner.h
@@ -6,8 +6,8 @@ class EnemySpawner : public GameObject
 {
 public:
 	//------------ HP関連 ------------//
-	int hp;						//現在のヒットポイント
-	int maxHP;					//最大HP
+	float hp;					//現在のヒットポイント
+	float maxHP;				//最大HP
 	float spawnInterval;		//敵の出現間隔
 	int spawnNum;				//敵の同時出現数
 	float spawnRandomPos;		//敵が出現する際の中心位置からのズレ
@@ -25,6 +25,6 @@ public:
 	void PopEnemy(const Vector3 position,float time);
 
 	//ダメージくらう
-	void Damage(uint32_t damage) { hp -= damage; }
+	void Damage(float damage) { hp -= damage; }
 };
