+++ Application/Object/Tower/Tower.h
@@ -5,8 +5,8 @@ class Tower : public GameObject
 {
 public:
 	//------------ HP関連 ------------//
-	uint32_t hp;				//現在のヒットポイント
-	uint32_t maxHP;				//最大HP
+	float hp;				//現在のヒットポイント
+	float maxHP;			//最大HP
 
 public:
 	Tower();
@@ -15,5 +15,5 @@ public:
 	void Draw()override;
 
 	//ダメージくらう
-	void Damage(uint32_t damage) { hp -= damage; }
+	void Damage(float damage) { hp -= damage; }
 };
